feat: reorganizing secretsManager folder and functions
